requirements.txt
torchvision
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
chromadb==0.4.18
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
open-clip-torch==2.23.0
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1
numpy<2.0.0